+++ src/Stack.Tests/Commands/PullRequests/CreatePullRequestsCommandHandlerTests.cs
@@ -30,14 +30,10 @@ public class CreatePullRequestsCommandHandlerTests(ITestOutputHelper testOutputH
         var gitHubClient = new TestGitHubRepositoryBuilder().Build();
         var stackConfig = Substitute.For<IStackConfig>();
         var inputProvider = Substitute.For<IInputProvider>();
-        var outputProvider = Substitute.For<IOutputProvider>();
+        var outputProvider = new TestOutputProvider(testOutputHelper);
         var fileOperations = new FileOperations();
         var gitClient = new GitClient(outputProvider, repo.GitClientSettings);
         var handler = new CreatePullRequestsCommandHandler(inputProvider, outputProvider, gitClient, gitHubClient, fileOperations, stackConfig);
-
-        outputProvider
-            .WhenForAnyArgs(o => o.Status(Arg.Any<string>(), Arg.Any<Action>()))
-            .Do(ci => ci.ArgAt<Action>(1)());
 
         var stacks = new List<Config.Stack>(
         [
@@ -99,14 +95,10 @@ A custom description
         var gitHubClient = new TestGitHubRepositoryBuilder().Build();
         var stackConfig = Substitute.For<IStackConfig>();
         var inputProvider = Substitute.For<IInputProvider>();
-        var outputProvider = Substitute.For<IOutputProvider>();
+        var outputProvider = new TestOutputProvider(testOutputHelper);
         var fileOperations = new FileOperations();
         var gitClient = new GitClient(outputProvider, repo.GitClientSettings);
         var handler = new CreatePullRequestsCommandHandler(inputProvider, outputProvider, gitClient, gitHubClient, fileOperations, stackConfig);
-
-        outputProvider
-            .WhenForAnyArgs(o => o.Status(Arg.Any<string>(), Arg.Any<Action>()))
-            .Do(ci => ci.ArgAt<Action>(1)());
 
         var stacks = new List<Config.Stack>(
         [
@@ -211,14 +203,10 @@ A custom description
         var gitHubClient = new TestGitHubRepositoryBuilder().Build();
         var stackConfig = Substitute.For<IStackConfig>();
         var inputProvider = Substitute.For<IInputProvider>();
-        var outputProvider = Substitute.For<IOutputProvider>();
+        var outputProvider = new TestOutputProvider(testOutputHelper);
         var fileOperations = new FileOperations();
         var gitClient = new GitClient(outputProvider, repo.GitClientSettings);
         var handler = new CreatePullRequestsCommandHandler(inputProvider, outputProvider, gitClient, gitHubClient, fileOperations, stackConfig);
-
-        outputProvider
-            .WhenForAnyArgs(o => o.Status(Arg.Any<string>(), Arg.Any<Action>()))
-            .Do(ci => ci.ArgAt<Action>(1)());
 
         var stacks = new List<Config.Stack>(
         [
@@ -273,14 +261,10 @@ A custom description
         var gitHubClient = new TestGitHubRepositoryBuilder().Build();
         var stackConfig = Substitute.For<IStackConfig>();
         var inputProvider = Substitute.For<IInputProvider>();
-        var outputProvider = Substitute.For<IOutputProvider>();
+        var outputProvider = new TestOutputProvider(testOutputHelper);
         var fileOperations = new FileOperations();
         var gitClient = new GitClient(outputProvider, repo.GitClientSettings);
         var handler = new CreatePullRequestsCommandHandler(inputProvider, outputProvider, gitClient, gitHubClient, fileOperations, stackConfig);
-
-        outputProvider
-            .WhenForAnyArgs(o => o.Status(Arg.Any<string>(), Arg.Any<Action>()))
-            .Do(ci => ci.ArgAt<Action>(1)());
 
         var stacks = new List<Config.Stack>(
         [
@@ -330,14 +314,10 @@ A custom description
         var gitHubClient = new TestGitHubRepositoryBuilder().Build();
         var stackConfig = Substitute.For<IStackConfig>();
         var inputProvider = Substitute.For<IInputProvider>();
-        var outputProvider = Substitute.For<IOutputProvider>();
+        var outputProvider = new TestOutputProvider(testOutputHelper);
         var fileOperations = new FileOperations();
         var gitClient = new GitClient(outputProvider, repo.GitClientSettings);
         var handler = new CreatePullRequestsCommandHandler(inputProvider, outputProvider, gitClient, gitHubClient, fileOperations, stackConfig);
-
-        outputProvider
-            .WhenForAnyArgs(o => o.Status(Arg.Any<string>(), Arg.Any<Action>()))
-            .Do(ci => ci.ArgAt<Action>(1)());
 
         var stacks = new List<Config.Stack>(
         [
@@ -375,14 +355,10 @@ A custom description
             .Build();
         var stackConfig = Substitute.For<IStackConfig>();
         var inputProvider = Substitute.For<IInputProvider>();
-        var outputProvider = Substitute.For<IOutputProvider>();
+        var outputProvider = new TestOutputProvider(testOutputHelper);
         var fileOperations = new FileOperations();
         var gitClient = new GitClient(outputProvider, repo.GitClientSettings);
         var handler = new CreatePullRequestsCommandHandler(inputProvider, outputProvider, gitClient, gitHubClient, fileOperations, stackConfig);
-
-        outputProvider
-            .WhenForAnyArgs(o => o.Status(Arg.Any<string>(), Arg.Any<Action>()))
-            .Do(ci => ci.ArgAt<Action>(1)());
 
         var stacks = new List<Config.Stack>(
         [
@@ -437,14 +413,10 @@ A custom description
         var gitHubClient = new TestGitHubRepositoryBuilder().Build();
         var stackConfig = Substitute.For<IStackConfig>();
         var inputProvider = Substitute.For<IInputProvider>();
-        var outputProvider = Substitute.For<IOutputProvider>();
+        var outputProvider = new TestOutputProvider(testOutputHelper);
         var fileOperations = new FileOperations();
         var gitClient = new GitClient(outputProvider, repo.GitClientSettings);
         var handler = new CreatePullRequestsCommandHandler(inputProvider, outputProvider, gitClient, gitHubClient, fileOperations, stackConfig);
-
-        outputProvider
-            .WhenForAnyArgs(o => o.Status(Arg.Any<string>(), Arg.Any<Action>()))
-            .Do(ci => ci.ArgAt<Action>(1)());
 
         Directory.CreateDirectory(Path.Join(repo.LocalDirectoryPath, ".github"));
         File.WriteAllText(Path.Join(repo.LocalDirectoryPath, ".github", "PULL_REQUEST_TEMPLATE.md"), "This is the PR template");
@@ -503,14 +475,10 @@ This is the PR template";
         var gitHubClient = new TestGitHubRepositoryBuilder().Build();
         var stackConfig = Substitute.For<IStackConfig>();
         var inputProvider = Substitute.For<IInputProvider>();
-        var outputProvider = Substitute.For<IOutputProvider>();
+        var outputProvider = new TestOutputProvider(testOutputHelper);
         var fileOperations = new FileOperations();
         var gitClient = new GitClient(outputProvider, repo.GitClientSettings);
         var handler = new CreatePullRequestsCommandHandler(inputProvider, outputProvider, gitClient, gitHubClient, fileOperations, stackConfig);
-
-        outputProvider
-            .WhenForAnyArgs(o => o.Status(Arg.Any<string>(), Arg.Any<Action>()))
-            .Do(ci => ci.ArgAt<Action>(1)());
 
         var stacks = new List<Config.Stack>(
         [
@@ -563,14 +531,10 @@ This is the PR template";
         var gitHubClient = new TestGitHubRepositoryBuilder().Build();
         var stackConfig = Substitute.For<IStackConfig>();
         var inputProvider = Substitute.For<IInputProvider>();
-        var outputProvider = Substitute.For<IOutputProvider>();
+        var outputProvider = new TestOutputProvider(testOutputHelper);
         var fileOperations = new FileOperations();
         var gitClient = new GitClient(outputProvider, repo.GitClientSettings);
         var handler = new CreatePullRequestsCommandHandler(inputProvider, outputProvider, gitClient, gitHubClient, fileOperations, stackConfig);
-
-        outputProvider
-            .WhenForAnyArgs(o => o.Status(Arg.Any<string>(), Arg.Any<Action>()))
-            .Do(ci => ci.ArgAt<Action>(1)());
 
         var stacks = new List<Config.Stack>(
         [
@@ -613,14 +577,10 @@ This is the PR template";
         var gitHubClient = new TestGitHubRepositoryBuilder().Build();
         var stackConfig = Substitute.For<IStackConfig>();
         var inputProvider = Substitute.For<IInputProvider>();
-        var outputProvider = Substitute.For<IOutputProvider>();
+        var outputProvider = new TestOutputProvider(testOutputHelper);
         var fileOperations = new FileOperations();
         var gitClient = new GitClient(outputProvider, repo.GitClientSettings);
         var handler = new CreatePullRequestsCommandHandler(inputProvider, outputProvider, gitClient, gitHubClient, fileOperations, stackConfig);
-
-        outputProvider
-            .WhenForAnyArgs(o => o.Status(Arg.Any<string>(), Arg.Any<Action>()))
-            .Do(ci => ci.ArgAt<Action>(1)());
 
         var stacks = new List<Config.Stack>(
         [
@@ -655,119 +615,5 @@ This is the PR template";
         };
 
         gitHubClient.PullRequests.Should().BeEquivalentTo(expectedPullRequests, ExcludeUnimportantPullRequestProperties);
-    }
-}
-
-public class TestGitHubPullRequestBuilder
-{
-    int _number = Some.Int();
-    string _title = Some.Name();
-    string _body = Some.Name();
-    string _state = GitHubPullRequestStates.Open;
-    Uri _url = Some.HttpsUri();
-    bool _draft = false;
-
-    public TestGitHubPullRequestBuilder WithNumber(int number)
-    {
-        _number = number;
-        return this;
-    }
-
-    public TestGitHubPullRequestBuilder WithTitle(string title)
-    {
-        _title = title;
-        return this;
-    }
-
-    public TestGitHubPullRequestBuilder WithBody(string body)
-    {
-        _body = body;
-        return this;
-    }
-
-    public TestGitHubPullRequestBuilder Merged()
-    {
-        _state = GitHubPullRequestStates.Merged;
-        return this;
-    }
-
-    public TestGitHubPullRequestBuilder Open()
-    {
-        _state = GitHubPullRequestStates.Open;
-        return this;
-    }
-
-    public TestGitHubPullRequestBuilder Closed()
-    {
-        _state = GitHubPullRequestStates.Closed;
-        return this;
-    }
-
-    public TestGitHubPullRequestBuilder WithUrl(Uri url)
-    {
-        _url = url;
-        return this;
-    }
-
-    public TestGitHubPullRequestBuilder AsDraft()
-    {
-        _draft = true;
-        return this;
-    }
-
-    public GitHubPullRequest Build()
-    {
-        return new GitHubPullRequest(_number, _title, _body, _state, _url, _draft);
-    }
-}
-
-public class TestGitHubRepositoryBuilder
-{
-    readonly Dictionary<string, GitHubPullRequest> _pullRequests = new();
-
-    public TestGitHubRepositoryBuilder WithPullRequest(string branch, Action<TestGitHubPullRequestBuilder> pullRequestBuilder)
-    {
-        var builder = new TestGitHubPullRequestBuilder();
-        pullRequestBuilder(builder);
-        _pullRequests.Add(branch, builder.Build());
-        return this;
-    }
-
-    public TestGitHubRepository Build()
-    {
-        return new TestGitHubRepository(_pullRequests);
-    }
-}
-
-public class TestGitHubRepository(Dictionary<string, GitHubPullRequest> PullRequests) : IGitHubClient
-{
-    public Dictionary<string, GitHubPullRequest> PullRequests { get; } = PullRequests;
-
-    public GitHubPullRequest CreatePullRequest(string headBranch, string baseBranch, string title, string bodyFilePath, bool draft)
-    {
-        var prBody = File.ReadAllText(bodyFilePath).Trim();
-        var pr = new GitHubPullRequest(Some.Int(), title, prBody, GitHubPullRequestStates.Open, Some.HttpsUri(), draft);
-        PullRequests.Add(headBranch, pr);
-        return pr;
-    }
-
-    public void EditPullRequest(int number, string body)
-    {
-        if (!PullRequests.Any(pr => pr.Value.Number == number))
-        {
-            throw new InvalidOperationException("Pull request not found.");
-        }
-
-        var pr = PullRequests.First(p => p.Value.Number == number);
-        PullRequests[pr.Key] = pr.Value with { Body = body };
-    }
-
-    public GitHubPullRequest? GetPullRequest(string branch)
-    {
-        return PullRequests.GetValueOrDefault(branch);
-    }
-
-    public void OpenPullRequest(GitHubPullRequest pullRequest)
-    {
     }
 }
